Chapter 5 completed

+++ Book_utilities/src/chapter5/BitwiseOperatorsTable.java
@@ -10,20 +10,20 @@ public class BitwiseOperatorsTable {
 
 		p = 0; q = 0;
 		System.out.print(p + "\t" + q + "\t");
-		System.out.print((p&q) + "\t" + (p|q) + "\t");
-		System.out.println((p^q) + "\t" + (~p));
+		System.out.print( ((byte) (p&q)) + "\t" + ((byte) (p|q)) + "\t");
+		System.out.println( ((byte) (p^q)) + "\t" + ((byte) (~p)));
 		p = 1; q = 0;
 		System.out.print(p + "\t" + q + "\t");
-		System.out.print((p&q) + "\t" + (p|q) + "\t");
-		System.out.println((p^q) + "\t" + (~p));
+		System.out.print( ((byte) (p&q)) + "\t" + ((byte) (p|q)) + "\t");
+		System.out.println( ((byte) (p^q)) + "\t" + ((byte) (~p)));
 		p = 0; q = 1;
 		System.out.print(p + "\t" + q + "\t");
-		System.out.print((p&q) + "\t" + (p|q) + "\t");
-		System.out.println((p^q) + "\t" + (~p));
+		System.out.print( ((byte) (p&q)) + "\t" + ((byte) (p|q)) + "\t");
+		System.out.println( ((byte) (p^q)) + "\t" + ((byte) (~p)));
 		p = 1; q = 1;
 		System.out.print(p + "\t" + q + "\t");
-		System.out.print((p&q) + "\t" + (p|q) + "\t");
-		System.out.println((p^q) + "\t" + (~p));
+		System.out.print( ((byte) (p&q)) + "\t" + ((byte) (p|q)) + "\t");
+		System.out.println( ((byte) (p^q)) + "\t" + ((byte) (~p)));
 
 
 	}
